+++ src/studyarc/gui/main_window.py
@@ -1,5 +1,7 @@
 import customtkinter as ctk
 
+from studyarc.config import colors
+from studyarc.config import constants
 
 from studyarc.gui.components.sidebar import Sidebar
 from studyarc.gui.components.content import Content
@@ -10,10 +12,14 @@ class MainWindow(ctk.CTk):
         super().__init__()
 
         self.title("StudyArc")
-        self.geometry("1000x700")
+        self.geometry(
+            f"{constants.WINDOW_WIDTH}x{constants.WINDOW_HEIGHT}"
+        )
+        self.configure(
+            fg_color=colors.BACKGROUND
+        )
+        self.sidebar = Sidebar(self)
+        self.sidebar.pack(side="left", fill="y")
 
-        sidebar = Sidebar(self)
-        sidebar.pack(side="left", fill="y")
-
-        content = Content(self)
-        content.pack(side="right", expand=True, fill="both")
+        self.content = Content(self)
+        self.content.pack(side="right", expand=True, fill="both")